requirements.txt
google-genai>=1.16.1
openlit==1.33.21
grpcio>=1.59.0 